--- src/main/java/com/example/timedeal/product/dto/request/CreateProductSaleRequest.java
@@ -1,14 +1,0 @@
-package com.example.timedeal.product.dto.request;
-
-import lombok.Data;
-
-import java.time.LocalDateTime;
-
-@Data
-public class CreateProductSaleRequest {
-    private LocalDateTime startedAt;
-
-    private LocalDateTime finishedAt;
-
-    private int discount;
-}
