+++ src/main/java/com/example/timedeal/timedeal/dto/response/UpdateTimeDealResponse.java
@@ -17,8 +17,6 @@ public class UpdateTimeDealResponse {
 
     private String name;
 
-    private int stock;
-
     private int discount;
 
     private LocalDateTime startedAt;
@@ -32,7 +30,6 @@ public class UpdateTimeDealResponse {
                 .id(timeDeal.getId())
                 .productId(timeDeal.getProduct().getId())
                 .name(timeDeal.getName())
-                .stock(timeDeal.getStock())
                 .discount(timeDeal.getDiscount())
                 .startedAt(timeDeal.getStartedAt())
                 .finishedAt(timeDeal.getFinishedAt())
