--- src/test/java/com/example/timedeal/mock/UserTest.java
@@ -1,14 +1,0 @@
-package com.example.timedeal.mock;
-
-import com.example.timedeal.user.domain.User;
-import com.example.timedeal.dto.user.request.CreateUserRequest;
-
-public class UserTest {
-    public static User get(CreateUserRequest request){
-        return User.builder()
-                .name(request.getName())
-                .password(request.getPassword())
-                .type(request.getType())
-                .build();
-    }
-}
